Kth Largest/Smallest Number
Problem Link - https://www.hackerrank.com/contests/snapdeal-codesprint/challenges/kth-largest-smallest-number/problem

You are given a sorted array  of  integers and the elements of the array are unique. There will be queries. For each query, you will be given three integers ,  and .

There are 2 types of queries:

For Type 0 queries (type = 0), you have to find the th smallest element among all the elements that are greater than . If the number of elements that are greater than  are less than , output should be .

For Type 1 queries (type = 1), you have to find the th largest element among all the elements that are smaller than . If the number of elements that are smaller than  are less than , output should be .

Input Constraints: 
 
 
 
 
 

Input Format

The first line of input will contain two integers:  and .
The next line will contain  integers. 
The following  lines each contain 3 integers: ,  and  respectively.

Output Format

Output  lines: the answer for each query in a separate line.

Sample Input

9 3
1 2 3 4 5 6 7 9 10
2 3 0
7 4 1
5 1 0
Sample Output

5
3
6
Explanation

For the 1st query, the elements greater than 2 are: 3, 4, 5, 6, 7, 9, 10.
The 3rd smallest among these is 5.

For the 2nd query, the elements smaller than 7 are: 1, 2, 3, 4, 5, 6.
The 4th largest among these is 3.

For the 3rd query, elements greater than 5 are: 6, 7, 9, 10.
The 1st smallest among these is 6.